Move more components into workspaces
