+++ client/web/src/enterprise/batches/batch-spec/execute/workspaces/ExecutionWorkspaces.module.scss
@@ -1,5 +1,6 @@
 .container {
     display: flex;
+    flex: 1;
     flex-direction: column;
     // Default of min-height: auto when this element is display: flex allows children to
     // set its height. We want to keep this fixed and have children overflow: auto
